Design array

Design_array contains item administration information at all time points in the Spatial Rotation Learning Program.

Design_array

Format

An array of dimension N-by-J-by-L, containing each subject's item administration.

Source

Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.

Details

The data object "Design_array" contains an array of dimension N-by-J-by-L indicating the items assigned (1/0) to each subject at each time point.

Author

Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper