Gibbs sampler for learning models

Runs MCMC to estimate parameters of any of the listed learning models.

MCMC_learning(
  Response_list,
  Q_list,
  model,
  test_order,
  Test_versions,
  chain_length,
  burn_in,
  Q_examinee = NULL,
  Latency_list = NULL,
  G_version = NA_integer_,
  theta_propose = 0,
  deltas_propose = NULL,
  R = NULL
)

Arguments

Response_list
A list of dichotomous item responses. t-th element is an N-by-Jt matrix of responses at time t.

Q_list
A list of Q-matrices. b-th element is a Jt-by-K Q-matrix for items in block b.

model
A charactor of the type of model fitted with the MCMC sampler, possible selections are "DINA_HO": Higher-Order Hidden Markov Diagnostic Classification Model with DINA responses; "DINA_HO_RT_joint": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and joint modeling of latent speed and learning ability; "DINA_HO_RT_sep": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and separate modeling of latent speed and learning ability; "rRUM_indept": Simple independent transition probability model with rRUM responses "NIDA_indept": Simple independent transition probability model with NIDA responses "DINA_FOHM": First Order Hidden Markov model with DINA responses

test_order
A matrix of the order of item blocks for each test version.

Test_versions
A vector of the test version of each learner.

chain_length
An int of the MCMC chain length.

burn_in
An int of the MCMC burn-in chain length.

Q_examinee
Optional. A list of the Q matrix for each learner. i-th element is a J-by-K Q-matrix for all items learner i was administered.

Latency_list
Optional. A list of the response times. t-th element is an N-by-Jt matrix of response times at time t.

G_version
Optional. An int of the type of covariate for increased fluency (1: G is dichotomous depending on whether all skills required for current item are mastered; 2: G cumulates practice effect on previous items using mastered skills; 3: G is a time block effect invariant across subjects with different attribute trajectories)

theta_propose
Optional. A scalar for the standard deviation of theta's proposal distribution in the MH sampling step.

deltas_propose
Optional. A vector for the band widths of each lambda's proposal distribution in the MH sampling step.

R
Optional. A reachability matrix for the hierarchical relationship between attributes.

Value

A list of parameter samples and Metropolis-Hastings acceptance rates (if applicable).

Examples

# \donttest{
output_FOHM = MCMC_learning(Y_real_list,Q_list,"DINA_FOHM",test_order,Test_versions,10000,5000)

#> 0
#> 1000
#> 2000
#> 3000
#> 4000
#> 5000
#> 6000
#> 7000
#> 8000
#> 9000

# }