Compute item pairwise odds ratio

Based on a response matrix, calculate the item pairwise odds-ratio according do (n11n00)/(n10n01), where nij is the number of people answering both item i and item j correctly

OddsRatio(N, J, Yt)

Arguments

N
An int of the sample size

J
An int of the number of items

Yt
An N-by-J response matrix

Value

A J-by-J upper-triangular matrix of the item pairwise odds ratios

Examples

# \donttest{
N = length(Test_versions)
Jt = nrow(Q_list[[1]])
K = ncol(Q_list[[1]])
T = nrow(test_order)
J = Jt*T
OddsRatio(N,Jt,Y_real_list[[1]])# }

#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,]    0 5.018707 1.779661 2.776068 2.096824 1.459009 1.908336 1.319590
#>  [2,]    0 0.000000 2.412364 4.487179 3.422177 1.772773 2.659884 2.070000
#>  [3,]    0 0.000000 0.000000 1.656774 1.978085 2.205882 1.365610 1.372881
#>  [4,]    0 0.000000 0.000000 0.000000 1.399892 2.834921 1.362193 1.440000
#>  [5,]    0 0.000000 0.000000 0.000000 0.000000 2.350859 2.296667 3.434117
#>  [6,]    0 0.000000 0.000000 0.000000 0.000000 0.000000 1.570304 1.766986
#>  [7,]    0 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 2.236145
#>  [8,]    0 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,]    0 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [10,]    0 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>            [,9]    [,10]
#>  [1,] 3.2507937 2.372730
#>  [2,] 2.1900000 2.215909
#>  [3,] 2.4734043 1.697626
#>  [4,] 1.8540305 1.766279
#>  [5,] 3.3515625 1.310606
#>  [6,] 0.9705882 1.984176
#>  [7,] 2.8505747 1.689375
#>  [8,] 1.6936364 1.437160
#>  [9,] 0.0000000 2.160088
#> [10,] 0.0000000 0.000000