Subjects' test version

Test_versions contains each subject's test module in the Spatial Rotation Learning Program.

Test_versions

Format

A vector of length N, containing each subject's assigned test module.

Source

Spatial Rotation Learning Experiment at UIUC between Fall 2015 and Spring 2016.

Details

The data object "Test_versions" contains a vector of length N indicating the test module assigned to each subject. Each test module consists of multiple item blocks with different orders over L time points. The order of item blocks corresponding to each test module is presented in the data object "Test_order".

See also

Author

Shiyu Wang, Yan Yang, Jeff Douglas, and Steve Culpepper