Gibbs sampler for learning models

Runs MCMC to estimate parameters of any of the listed learning models.

hmcdm(
  Response,
  Q_matrix,
  model,
  Design_array = NULL,
  Test_order = NULL,
  Test_versions = NULL,
  chain_length = 100L,
  burn_in = 50L,
  G_version = NA_integer_,
  theta_propose = 0,
  Latency_array = NULL,
  deltas_propose = NULL,
  R = NULL
)

Arguments

Response

An array of dichotomous item responses. t-th slice is an N-by-J matrix of responses at time t.

Q_matrix

A J-by-K Q-matrix.

model

A charactor of the type of model fitted with the MCMC sampler, possible selections are "DINA_HO": Higher-Order Hidden Markov Diagnostic Classification Model with DINA responses; "DINA_HO_RT_joint": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and joint modeling of latent speed and learning ability; "DINA_HO_RT_sep": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and separate modeling of latent speed and learning ability; "rRUM_indept": Simple independent transition probability model with rRUM responses "NIDA_indept": Simple independent transition probability model with NIDA responses "DINA_FOHM": First Order Hidden Markov model with DINA responses

Design_array

An array of dimension N-by-J-by-L indicating the items assigned (1/0) to each subject at each time point. Either 'Design_array' or both 'Test_order' & 'Test_versions' need to be provided to run HMCDM.

Test_order

Optional. A matrix of the order of item blocks for each test version.

Test_versions

Optional. A vector of the test version of each learner.

chain_length

An int of the MCMC chain length.

burn_in

An int of the MCMC burn-in chain length.

G_version

Optional. An int of the type of covariate for increased fluency (1: G is dichotomous depending on whether all skills required for current item are mastered; 2: G cumulates practice effect on previous items using mastered skills; 3: G is a time block effect invariant across subjects with different attribute trajectories)

theta_propose

Optional. A scalar for the standard deviation of theta's proposal distribution in the MH sampling step.

Latency_array

Optional. A array of the response times. t-th slice is an N-by-J matrix of response times at time t.

deltas_propose

Optional. A vector for the band widths of each lambda's proposal distribution in the MH sampling step.

R

Optional. A reachability matrix for the hierarchical relationship between attributes.

Value

A list of parameter samples and Metropolis-Hastings acceptance rates (if applicable).

Author

Susu Zhang

Examples

# \donttest{
output_FOHM = hmcdm(Y_real_array, Q_matrix, "DINA_FOHM", Design_array, 100, 30)
#> 0
# }