Obtain learning model point estimates

Obtain EAPs of continuous parameters and EAP or MAP of the attribute trajectory estimates under the CDM learning models based on the MCMC output

point_estimates_learning(output, model, N, Jt, K, T, alpha_EAP = TRUE)

Arguments

output
A list of MCMC outputs, obtained from the MCMC_learning function

model
A charactor of the type of model fitted with the MCMC sampler, possible selections are "DINA_HO": Higher-Order Hidden Markov Diagnostic Classification Model with DINA responses; "DINA_HO_RT_joint": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and joint modeling of latent speed and learning ability; "DINA_HO_RT_sep": Higher-Order Hidden Markov DCM with DINA responses, log-Normal response times, and separate modeling of latent speed and learning ability; "rRUM_indept": Simple independent transition probability model with rRUM responses "NIDA_indept": Simple independent transition probability model with NIDA responses "DINA_FOHM": First Order Hidden Markov model with DINA responses

N
An int of number of subjects

Jt
An int of number of items in each block

K
An int of number of skills

T
An int of number of time points

alpha_EAP
A boolean operator (T/F) of whether to use EAP for alphas (if F: use most likely trajectory (MAP) for alphas)

Value

A list of point estimates of model parameters

Examples

# \donttest{
N = length(Test_versions)
Jt = nrow(Q_list[[1]])
K = ncol(Q_list[[1]])
T = nrow(test_order)
J = Jt*T
output_FOHM = MCMC_learning(Y_real_list,Q_list,"DINA_FOHM",test_order,Test_versions,10000,5000)

#> 0
#> 1000
#> 2000
#> 3000
#> 4000
#> 5000
#> 6000
#> 7000
#> 8000
#> 9000

point_estimates = point_estimates_learning(output_FOHM,"DINA_FOHM",N,Jt,K,T,alpha_EAP = T)
# }