Generate attribute trajectories under the first order hidden Markov model

Based on the initial attribute patterns and probability of transitioning between different patterns, create cube of attribute patterns of all subjects across time.

simulate_alphas_FOHM(Omega, alpha0s, L)

Arguments

Omega

A 2^K-by-2^K matrix of transition probabilities from row pattern to column pattern

alpha0s

An N-by-K matrix of subjects' initial attribute patterns.

L

An int of number of time points

Value

An N-by-K-by-L array of attribute patterns of subjects at each time point.

Examples

N = length(Test_versions)
J = nrow(Q_matrix)
K = ncol(Q_matrix)
L = nrow(Test_order)
Jt = J/L
TP <- TPmat(K)
Omega_true <- rOmega(TP)
class_0 <- sample(1:2^K, N, replace = L)
Alphas_0 <- matrix(0,N,K)
for(i in 1:N){
  Alphas_0[i,] <- inv_bijectionvector(K,(class_0[i]-1))
}
Alphas <- simulate_alphas_FOHM(Omega_true, Alphas_0,L)