Generate attribute trajectories under the simple independent-attribute learning model

Based on the initial attribute patterns and probability of transitioning from 0 to 1 on each attribute, create cube of attribute patterns of all subjects across time. Transitions on different skills are regarded as independent.

simulate_alphas_indept(taus, alpha0s, T, R)

Arguments

taus
A length K vector of transition probabilities from 0 to 1 on each skill

alpha0s
An N-by-K matrix of subjects' initial attribute patterns.

T
An int of number of time points

R
A K-by-K dichotomous reachability matrix indicating the attribute hierarchies. The k,k'th entry of R is 1 if k' is prereq to k.

Value

An N-by-K-by-T array of attribute patterns of subjects at each time point.

Examples

N = length(Test_versions)
Jt = nrow(Q_list[[1]])
K = ncol(Q_list[[1]])
T = nrow(test_order)
J = Jt*T
tau <- numeric(K)
for(k in 1:K){
  tau[k] <- runif(1,.2,.6)
}
R = matrix(0,K,K)
# Initial alphas
p_mastery <- c(.5,.5,.4,.4)
Alphas_0 <- matrix(0,N,K)
for(i in 1:N){
  for(k in 1:K){
    prereqs <- which(R[k,]==1)
    if(length(prereqs)==0){
      Alphas_0[i,k] <- rbinom(1,1,p_mastery[k])
    }
    if(length(prereqs)>0){
      Alphas_0[i,k] <- prod(Alphas_0[i,prereqs])*rbinom(1,1,p_mastery)
    }
  }
}
Alphas <- simulate_alphas_indept(tau,Alphas_0,T,R)