Summarizing Hidden Markov Cognitive Diagnosis Model Fits

summary method for class "hmcdm" or "summary.hmcdm".

# S3 method for class 'summary.hmcdm'
print(x, ...)

# S3 method for class 'hmcdm'
summary(object, ...)

Arguments

x

an object of class "hmcdm.summary".

...

further arguments passed to or from other methods.

object

a fitted model object of class "hmcdm".

Value

The function summary.hmcdm computes and returns a list of point estimates of model parameters and model fit measures including DIC and PPP-values.

See also

Examples

# \donttest{
output_FOHM = hmcdm(Y_real_array,Q_matrix,"DINA_FOHM",Design_array,1000,500)
#> 0
summary(output_FOHM)
#> 
#> Model: DINA_FOHM 
#> 
#> Item Parameters:
#>   ss_EAP gs_EAP
#>  0.04157 0.7409
#>  0.04786 0.6545
#>  0.16422 0.5187
#>  0.07325 0.4378
#>  0.19364 0.4465
#>    ... 45 more items
#> 
#> Transition Parameters:
#>  [1] 0.732477 0.015800 0.029361 0.025451 0.008678 0.019483 0.013362 0.011147
#>  [9] 0.047119 0.011093 0.018802 0.009355 0.009795 0.013595 0.020006 0.014477
#>    ... 15 more rows
#> 
#> Class Probabilities:
#>       pis_EAP
#> 0000 0.111669
#> 0001 0.004885
#> 0010 0.027591
#> 0011 0.025328
#> 0100 0.074747
#>    ... 11 more classes
#> 
#> Deviance Information Criterion (DIC): 17885.95 
#> 
#> Posterior Predictive P-value (PPP):
#> M1:  0.57
#> M2:  0.49
#> total scores:  0.6319
# }