Compute item pairwise odds ratio

Based on a response matrix, calculate the item pairwise odds-ratio according do (n11n00)/(n10n01), where nij is the number of people answering both item i and item j correctly

OddsRatio(N, J, Yt)

Arguments

N

An int of the sample size

J

An int of the number of items

Yt

An N-by-J response matrix

Value

A J-by-J upper-triangular matrix of the item pairwise odds ratios

Examples

# \donttest{
N = dim(Y_real_array)[1]
J = nrow(Q_matrix)
OddsRatio(N,J,Y_real_array[,,1])# }
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [2,]    0    0   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [3,]    0    0    0   NA   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [4,]    0    0    0    0   NA   NA   NA   NA   NA    NA    NA    NA    NA
#>  [5,]    0    0    0    0    0   NA   NA   NA   NA    NA    NA    NA    NA
#>  [6,]    0    0    0    0    0    0   NA   NA   NA    NA    NA    NA    NA
#>  [7,]    0    0    0    0    0    0    0   NA   NA    NA    NA    NA    NA
#>  [8,]    0    0    0    0    0    0    0    0   NA    NA    NA    NA    NA
#>  [9,]    0    0    0    0    0    0    0    0    0    NA    NA    NA    NA
#> [10,]    0    0    0    0    0    0    0    0    0     0    NA    NA    NA
#> [11,]    0    0    0    0    0    0    0    0    0     0     0    NA    NA
#> [12,]    0    0    0    0    0    0    0    0    0     0     0     0    NA
#> [13,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [14,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [15,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [16,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [17,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [18,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [19,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [20,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [21,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [22,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [23,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [24,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [25,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [26,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [27,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [28,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [29,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [30,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [31,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [32,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [33,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [34,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [35,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [36,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [37,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [38,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [39,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [40,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [41,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [42,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [43,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [44,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [45,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [46,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [47,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [48,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [49,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#> [50,]    0    0    0    0    0    0    0    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [7,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [8,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [14,]     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]     0     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]     0     0     0     0     0    NA    NA    NA    NA    NA    NA    NA
#> [19,]     0     0     0     0     0     0    NA    NA    NA    NA    NA    NA
#> [20,]     0     0     0     0     0     0     0    NA    NA    NA    NA    NA
#> [21,]     0     0     0     0     0     0     0     0    NA    NA    NA    NA
#> [22,]     0     0     0     0     0     0     0     0     0    NA    NA    NA
#> [23,]     0     0     0     0     0     0     0     0     0     0    NA    NA
#> [24,]     0     0     0     0     0     0     0     0     0     0     0    NA
#> [25,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0     0     0     0     0     0
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [7,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [8,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [23,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [24,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [25,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [26,]     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [27,]     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [28,]     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [29,]     0     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA
#> [30,]     0     0     0     0     0    NA    NA    NA    NA    NA    NA    NA
#> [31,]     0     0     0     0     0     0    NA    NA    NA    NA    NA    NA
#> [32,]     0     0     0     0     0     0     0    NA    NA    NA    NA    NA
#> [33,]     0     0     0     0     0     0     0     0    NA    NA    NA    NA
#> [34,]     0     0     0     0     0     0     0     0     0    NA    NA    NA
#> [35,]     0     0     0     0     0     0     0     0     0     0    NA    NA
#> [36,]     0     0     0     0     0     0     0     0     0     0     0    NA
#> [37,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0     0     0     0     0     0
#>       [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#>  [1,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [2,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [3,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [4,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [5,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [6,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [7,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [8,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [22,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [23,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [24,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [25,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [26,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [27,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [28,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [29,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [30,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [31,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [32,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [33,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [34,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [35,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [36,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [37,]    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [38,]     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [39,]     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [40,]     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [41,]     0     0     0     0    NA    NA    NA    NA    NA    NA    NA    NA
#> [42,]     0     0     0     0     0    NA    NA    NA    NA    NA    NA    NA
#> [43,]     0     0     0     0     0     0    NA    NA    NA    NA    NA    NA
#> [44,]     0     0     0     0     0     0     0    NA    NA    NA    NA    NA
#> [45,]     0     0     0     0     0     0     0     0    NA    NA    NA    NA
#> [46,]     0     0     0     0     0     0     0     0     0    NA    NA    NA
#> [47,]     0     0     0     0     0     0     0     0     0     0    NA    NA
#> [48,]     0     0     0     0     0     0     0     0     0     0     0    NA
#> [49,]     0     0     0     0     0     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0     0     0     0     0     0
#>       [,50]
#>  [1,]    NA
#>  [2,]    NA
#>  [3,]    NA
#>  [4,]    NA
#>  [5,]    NA
#>  [6,]    NA
#>  [7,]    NA
#>  [8,]    NA
#>  [9,]    NA
#> [10,]    NA
#> [11,]    NA
#> [12,]    NA
#> [13,]    NA
#> [14,]    NA
#> [15,]    NA
#> [16,]    NA
#> [17,]    NA
#> [18,]    NA
#> [19,]    NA
#> [20,]    NA
#> [21,]    NA
#> [22,]    NA
#> [23,]    NA
#> [24,]    NA
#> [25,]    NA
#> [26,]    NA
#> [27,]    NA
#> [28,]    NA
#> [29,]    NA
#> [30,]    NA
#> [31,]    NA
#> [32,]    NA
#> [33,]    NA
#> [34,]    NA
#> [35,]    NA
#> [36,]    NA
#> [37,]    NA
#> [38,]    NA
#> [39,]    NA
#> [40,]    NA
#> [41,]    NA
#> [42,]    NA
#> [43,]    NA
#> [44,]    NA
#> [45,]    NA
#> [46,]    NA
#> [47,]    NA
#> [48,]    NA
#> [49,]    NA
#> [50,]     0