Simulate DINA model responses (entire cube)

Simulate a cube of DINA responses for all persons on items across all time points

simDINA(alphas, itempars, ETA, test_order, Test_versions)

Arguments

alphas
An N-by-K-by-T array of attribute patterns of all persons across T time points

itempars
A J-by-2-by-T cube of item parameters (slipping: 1st col, guessin: 2nd col) across item blocks

ETA
A J-by-2^K-by-T array of ideal responses across all item blocks, with each slice generated with ETAmat function

test_order
A N_versions-by-T matrix indicating which block of items were administered to examinees with specific test version.

Test_versions
A length N vector of the test version of each examinee

Value

An array of DINA item responses of examinees across all time points

Examples

N = length(Test_versions)
Jt = nrow(Q_list[[1]])
K = ncol(Q_list[[1]])
T = nrow(test_order)
J = Jt*T
itempars_true <- array(runif(Jt*2*T,.1,.2), dim = c(Jt,2,T))

ETAs <- array(NA,dim = c(Jt,2^K,T))
for(t in 1:T){
  ETAs[,,t] <- ETAmat(K,Jt,Q_list[[t]])
}
class_0 <- sample(1:2^K, N, replace = T)
Alphas_0 <- matrix(0,N,K)
mu_thetatau = c(0,0)
Sig_thetatau = rbind(c(1.8^2,.4*.5*1.8),c(.4*.5*1.8,.25))
Z = matrix(rnorm(N*2),N,2)
thetatau_true = Z%*%chol(Sig_thetatau)
thetas_true = thetatau_true[,1]
taus_true = thetatau_true[,2]
G_version = 3
phi_true = 0.8
for(i in 1:N){
  Alphas_0[i,] <- inv_bijectionvector(K,(class_0[i]-1))
}
lambdas_true <- c(-2, .4, .055)
Alphas <- simulate_alphas_HO_joint(lambdas_true,thetas_true,Alphas_0,Q_examinee,T,Jt)
Y_sim <- simDINA(Alphas,itempars_true,ETAs,test_order,Test_versions)