Summarizing Hidden Markov Cognitive Diagnosis Model Fits

summary method for class "hmcdm" or "summary.hmcdm".

# S3 method for summary.hmcdm
print(x, ...)

# S3 method for hmcdm
summary(object, ...)

Arguments

x

an object of class "hmcdm.summary".

...

further arguments passed to or from other methods.

object

a fitted model object of class "hmcdm".

Value

The function summary.hmcdm computes and returns a list of point estimates of model parameters and model fit measures including DIC and PPP-values.

See also

Examples

# \donttest{
output_FOHM = hmcdm(Y_real_array,Q_matrix,"DINA_FOHM",Design_array,1000,500)
#> 0
summary(output_FOHM)
#> 
#> Model: DINA_FOHM 
#> 
#> Item Parameters:
#>   ss_EAP gs_EAP
#>  0.03857 0.7087
#>  0.05056 0.6874
#>  0.16675 0.4526
#>  0.06765 0.5243
#>  0.19003 0.4132
#>    ... 45 more items
#> 
#> Transition Parameters:
#>  [1] 0.738655 0.009169 0.017145 0.013539 0.025876 0.016524 0.015503 0.011819
#>  [9] 0.054643 0.013102 0.015012 0.012754 0.012391 0.016374 0.015644 0.011851
#>    ... 15 more rows
#> 
#> Class Probabilities:
#>       pis_EAP
#> 0000 0.165350
#> 0001 0.011494
#> 0010 0.006487
#> 0011 0.013762
#> 0100 0.023602
#>    ... 11 more classes
#> 
#> Deviance Information Criterion (DIC): 18059.24 
#> 
#> Posterior Predictive P-value (PPP):
#> M1: 0.542
#> M2:  0.49
#> total scores:  0.6301
# }